+++ src/test/java/faang/school/postservice/controller/TestLikeController.java
@@ -1,5 +1,6 @@
 package faang.school.postservice.controller;
 
+import faang.school.postservice.controller.like.LikeController;
 import faang.school.postservice.dto.like.LikeDto;
 import faang.school.postservice.service.LikeService;
 import org.junit.jupiter.api.BeforeEach;
