+++ src/test/java/faang/school/postservice/service/LikeServiceTest.java
@@ -1,11 +1,17 @@
 package faang.school.postservice.service;
 
 import faang.school.postservice.client.UserServiceClient;
+import faang.school.postservice.dto.like.LikeDto;
 import faang.school.postservice.dto.user.UserDto;
+import faang.school.postservice.mapper.LikeMapper;
 import faang.school.postservice.model.Comment;
 import faang.school.postservice.model.Like;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.repository.LikeRepository;
+import faang.school.postservice.validator.LikeServiceValidator;
+import feign.FeignException;
+import feign.Request;
+import feign.RequestTemplate;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -16,8 +22,12 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 import static org.mockito.Mockito.when;
 
@@ -33,23 +43,61 @@ public class LikeServiceTest {
     @Mock
     private UserServiceClient userServiceClient;
 
+    @Mock
+    private LikeServiceValidator likeServiceValidator;
+    @Mock
+    private PostService postService;
+    @Mock
+    private CommentService commentService;
+    @Mock
+    private LikeMapper likeMapper;
+
     @InjectMocks
     private LikeService likeService;
 
     private final List<Like> likes = new ArrayList<>();
     private final List<UserDto> usersDto = new ArrayList<>();
+    private long postId;
+    private long userId;
+    private long commentId;
+    private LikeDto likeDtoPost;
+    private LikeDto likeDtoComment;
+    private Post post;
+    private Comment comment;
+    private LikeDto likeDto;
 
     @BeforeEach
     public void setup() {
+        postId = 1;
+        userId = 1;
+        commentId = 1;
+
+        likeDtoPost = LikeDto.builder()
+                .id(1L)
+                .userId(1L)
+                .postId(1L)
+                .build();
+
+        likeDtoComment = LikeDto.builder()
+                .id(1L)
+                .userId(1)
+                .commentId(1)
+                .build();
+
+        likeDto = new LikeDto();
+        likeDto.setId(1L);
+        likeDto.setUserId(1L);
 
         likeService.setBATCH_SIZE(BATCH_SIZE);
 
-        Comment comment = Comment.builder()
+        comment = Comment.builder()
                 .id(2L)
+                .likes(new ArrayList<>(Arrays.asList(new Like(), new Like())))
                 .build();
 
-        Post post = Post.builder()
+        post = Post.builder()
                 .id(1L)
+                .likes(new ArrayList<>(Arrays.asList(new Like(), new Like())))
                 .build();
 
         for (long i = 1; i <= 5; i++) {
@@ -98,5 +146,74 @@ public class LikeServiceTest {
 
         verify(likeRepository, times(1)).findByCommentId(anyLong());
         verify(userServiceClient, times(1)).getUsersByIds(anyList());
+    }
+
+    @DisplayName("Когда метод по добавлению лайка к посту отработал")
+    @Test
+    public void testAddLikeToPostWhenValid() {
+        UserDto userDto = new UserDto(1L, "name", "email@google.com");
+        Like like = new Like();
+
+        when(postService.getPost(likeDtoPost.getPostId())).thenReturn(post);
+        when(userServiceClient.getUser(likeDtoPost.getUserId())).thenReturn(userDto);
+        when(likeRepository.findByPostIdAndUserId(post.getId(), userDto.getId())).thenReturn(Optional.empty());
+        when(likeMapper.toEntity(likeDtoPost)).thenReturn(like);
+        when(likeMapper.toLikeDto(like)).thenReturn(likeDto);
+
+        likeService.addLikeToPost(likeDtoPost);
+
+        verify(likeServiceValidator, times(1)).checkDuplicateLike(Optional.empty());
+        verify(likeRepository, times(1)).save(like);
+    }
+
+    @DisplayName("Когда метод по удалению лайка с поста отработал")
+    @Test
+    public void testDeleteLikeFromPostWhenValid() {
+        Like like = new Like();
+        like.setId(1);
+        Optional<Like> optionalLike = Optional.of(like);
+        post.setLikes(Arrays.asList(like, new Like()));
+
+        when(likeRepository.findByPostIdAndUserId(postId, userId)).thenReturn(optionalLike);
+        when(postService.getPost(postId)).thenReturn(post);
+
+        likeService.deleteLikeFromPost(postId, userId);
+        verify(likeRepository, times(1)).deleteByPostIdAndUserId(postId, userId);
+        verify(likeServiceValidator, times(1)).checkAvailabilityLike(optionalLike);
+    }
+
+    @DisplayName("Когда метод по добавлению лайка на комментарий отработал")
+    @Test
+    public void testAddLikeToCommentWhenValid() {
+        Like like = new Like();
+        UserDto userDto = new UserDto(userId, "name", "email@google.com");
+
+        when(commentService.getComment(likeDtoComment.getCommentId())).thenReturn(comment);
+        when(userServiceClient.getUser(likeDtoComment.getUserId())).thenReturn(userDto);
+        when(likeRepository.findByCommentIdAndUserId(comment.getId(), userDto.getId())).thenReturn(Optional.empty());
+        when(likeMapper.toEntity(likeDtoComment)).thenReturn(like);
+        when(likeMapper.toLikeDto(like)).thenReturn(likeDto);
+
+        likeService.addLikeToComment(likeDtoComment);
+
+        verify(likeServiceValidator, times(1)).checkDuplicateLike(Optional.empty());
+        verify(likeRepository, times(1)).save(like);
+    }
+
+    @DisplayName("Когда метод по удалению лайка с комментария отработал")
+    @Test
+    public void testDeleteLikeFromCommentWhenValid() {
+        Like like = new Like();
+        like.setId(1);
+        Optional<Like> optionalLike = Optional.of(like);
+        List<Like> likes = Arrays.asList(like);
+        comment.setLikes(likes);
+
+        when(likeRepository.findByCommentIdAndUserId(commentId, userId)).thenReturn(optionalLike);
+        when(commentService.getComment(commentId)).thenReturn(comment);
+
+        likeService.deleteLikeFromComment(commentId, userId);
+        verify(likeRepository, times(1)).deleteByCommentIdAndUserId(commentId, userId);
+        verify(likeServiceValidator, times(1)).checkAvailabilityLike(optionalLike);
     }
 }
